+++ src/Sharpen.VisualStudioExtension/Properties/AssemblyInfo.cs
@@ -3,7 +3,7 @@ internal static partial class AssemblyDescription
 {
     public const string Title = "Sharpen for Visual Studio";
 
-    public const string Description = "Sharpen is a Visual Studio extension that intelligently introduces new C# language features into your existing codebase. It makes your code smaller, simpler, faster, more expressive and more readable, in one word - sharper.";
+    public const string Description = "Sharpen is a Visual Studio extension that intelligently modernizes your existing C# codebase. It makes your code more expressive, more readable, more robust, and more performant, in one word - sharper.";
 
     public const string InformationalVersion = Version;
 }