+++ src/Sharpen.VisualStudioExtension/UserInteraction.cs
@@ -4,14 +4,19 @@ namespace Sharpen.VisualStudioExtension
 {
     internal static class UserInteraction
     {
-        public static void ShowInformation(string message)
+        public static void ShowPlainMessage(string message, string caption = "Sharpen")
         {
-            MessageBox.Show(message, "Sharpen", MessageBoxButton.OK, MessageBoxImage.Information);
+            MessageBox.Show(message, caption, MessageBoxButton.OK, MessageBoxImage.None);
         }
 
-        public static void ShowError(string message)
+        public static void ShowInformation(string message, string caption = "Sharpen")
         {
-            MessageBox.Show(message, "Sharpen", MessageBoxButton.OK, MessageBoxImage.Error);
+            MessageBox.Show(message, caption, MessageBoxButton.OK, MessageBoxImage.Information);
+        }
+
+        public static void ShowError(string message, string caption = "Sharpen")
+        {
+            MessageBox.Show(message, caption, MessageBoxButton.OK, MessageBoxImage.Error);
         }
     }
 }